Review
Overall this looks fine, my linter had some mentions about the order some of the 'member-access' statements needed to be in so I've changed that. Also I've updated some of the typing and (empty) checks to be more robust.

Getting and Setting of properties is not really a common practice. It is in the other typed languages but Typescript not so much. 
Also I don't really like the global 'lc' variable but I understand the use-case.